本地代码更新，新增一些功能。同事增加多线程的一些学习代码

+++ src/main/java/com/example/mapper/UserTjMapper.java
@@ -3,7 +3,9 @@ package com.example.mapper;
 import java.util.Map;
 
 public interface UserTjMapper {
-	
-        Map<String, Integer>  getAllByUserSex();
+        
+	    
+        Map<String, Long>  getAllByUserSex();
+        
 
 }
